pyproject.toml
[project]
name = "cephplaypen"
version = "0.1.0"
description = "Muckin abaht wiv CEPH"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiobotocore>=3.3.0",
    "backports-zstd>=1.4.0",
    "boto3>=1.42.63",
    "matplotlib>=3.10.8",
    "pymongo>=4.16.0",
    "urllib3[socks]>=2.6.3",
]

[dependency-groups]
dev = [
    "ipython>=9.11.0",
]
